check deep scan status consistency with individual scans
